Added WordsAPI functionality. Major bugs found.
WordsAPI functionality added to Utility.swift file. Currently, it has to
implement a 2 second sleep function due to async in order to retrieve
the data from the web. Working on figuring out other ways to manage
this.

In terms of bugs found (pretty sure this was fixed last time I checked)
but the current issue is that when words that are NOT in the database
are entered into the text field (to be translated) these words cause the
program to crash when attributedString (which is an optional) is being
assigned to attributedText to be displayed on the UI screen.
Last I remember, any words that didn't exist were highlighted in red and
the app would keep the user on the same input string until the user
typed in a valid input. This is a serious issue we should definitely
look into.