Displays full input when showing errors
Currently displays the entire input when showing errors and highlights
words NOT in the database in red, however sometimes highlights the
incorrect words (e.g. 'the man ate the dog') and it highlights 'man ate'
even though 'man' exists in the DB.

Also changed the implementation of 'getSentenceToWords' to take in an
extra optional parameter (default value = true) which decides whether it
returns the input string as an array of words directly, or actually
filters out the padding words and only returns relevant words to be
searched for in the database.